+++ js/constants.js
@@ -1,0 +1,38 @@
+const NAMES = [
+  'Сильвер',
+  'Рован',
+  'Сильвия',
+  'Йользе',
+  'Сатрана',
+  'Шелена',
+  'Торвальд',
+  'Иоланда',
+  'Корвин',
+  'Термина'
+];
+
+const AVATARS = ['1', '2', '3', '4', '5', '6'];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTIONS = [
+  'Путешествия',
+  'Домашние любимцы',
+  'Портреты',
+  'Пейзажи',
+  'Подводный мир',
+  'Дикая природа',
+  'Урбанистика',
+  'Документалистика',
+];
+
+const AMOUNT_OF_PHOTOS = 25;
+
+export {NAMES, AVATARS, MESSAGES, DESCRIPTIONS, AMOUNT_OF_PHOTOS};
